Material enum
the material is now an enumerated value instead of being a string. should help with readability and possibly typing errors since string comparison can lead to unforeseen bugs.

To use: there is an example in the Material section of user inputs but in general its just Material._<insert material here>_ all the possible materials are stored under the enumeration of the class Material in Material.m